+++ src/routes/(client)/threads/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { readableTime, colorHash } from '$lib/utils/formatting'
-	import { userState } from '../../userState'
+	import { userState } from '../../../userState'
 
 	let data: Thread[] | Promise<Thread[]> = []
 	const loadThreads = () =>